refactored CopyCode component
