+++ packages/api-explorer/src/block-types/CodeElement.jsx
@@ -1,7 +1,7 @@
 const React = require('react');
 const PropTypes = require('prop-types');
 
-const CopyCode = require('../CopyCode');
+const CopyCode = require('../components/CopyCode');
 const syntaxHighlighter = require('@readme/syntax-highlighter');
 
 /*
