+++ packages/api-explorer/src/components/CopyCode/index.jsx
@@ -1,3 +1,5 @@
+import './style.css'
+
 const React = require('react');
 const PropTypes = require('prop-types');
 const { CopyToClipboard } = require('react-copy-to-clipboard');
@@ -5,7 +7,6 @@ const { CopyToClipboard } = require('react-copy-to-clipboard');
 class CopyCode extends React.Component {
   constructor(props) {
     super(props);
-
     this.state = {
       copied: false,
     };
@@ -15,7 +16,6 @@ class CopyCode extends React.Component {
 
   componentDidUpdate(prevProps) {
     const code = typeof prevProps.code === 'function' ? prevProps.code() : prevProps.code;
-
     if (code !== this.state.code) {
       // eslint-disable-next-line react/no-did-update-set-state
       this.setState({ code });
@@ -30,11 +30,19 @@ class CopyCode extends React.Component {
   }
 
   render() {
+    // const buttonStyle = {
+    //   textTransform: 'uppercase',
+    //   background: 'transparent',
+    //   color: '#aaaaaa',
+    //   cursor: 'pointer',
+    //   border: '0px',
+    // }
+
     return (
-      <CopyToClipboard text={this.state.code} onCopy={this.onCopy}>
-        <button className="copy-code-button main_background">
+      <CopyToClipboard class="mia-ctc-button" text={this.state.code} onCopy={this.onCopy}>
+        <span>
           {this.state.copied ? <span>Copied</span> : <span>Copy</span>}
-        </button>
+        </span>
       </CopyToClipboard>
     );
   }
